+++ Krowi_AchievementFilter/Data/ExportedNextPatchAchievements.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2021-07-19 23-42-00 ]] --
+-- [[ Exported at 2021-07-26 01-29-24 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --